requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
idna==3.10
iniconfig==2.1.0
packaging==24.2
pillow==11.2.0
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
requests==2.32.3
urllib3==2.3.0
PyPDF2>=2.0.0
pycryptodome>=3.14.0
